Minimal versions of default exposure classes required for supplychain workflow.
For representative results in the supplychain workflow, these will need rework. ESPECIALLY IMPORTANT: currently, only the manufacturing exposure class is developed, the mining and utilitzitites classes use the same underlying data as manufacturing the manufacturing class. This is so that supplychain workflow can already be used and tested. In future, better/more accurate exposures for the various industry sectors have to be developed.