Speed up Impact's select by centroid
The existing method for subsetting an Impact by centroid (list
comprehension) was slow for large Impact objects with large selections.

The new method converts the method's coordinate ndarrays to pandas data
frames, does a join, and records the indices of common values. It's a
lot faster.

There may be an even better way, but this takes large calculations down
from ~40 seconds to < 1 sec which is all I need.