LitPop: skip countries without pixels
For coarse resolutions (e.g. 300 arc-seconds), small countries like Monaco don't get
any pixel in the grid. This will make LitPop crash. Hopefully, this will be fixed
in the reimplementation anyways. This is just a hot-fix to make the old
implementation more useable as long as the new one isn't available.